Pre-generate header/doc files based on isa-spec.yaml (#65)
When isa-spec.yaml is updated, pre-generated files must be recreated
through isa-spec-gen target.

```
cmake --build . --config Release --target isa-spec-gen
```

This PR moves the generation from src/cpp/aiebu/src/CMakeLists.txt
into
- specification/CMakeLists.txt and
- specification/{aie2,aie2ps}/CMakeLists.txt

The pre-generated files are created in the source tree where they are
checked in.

The isa-spec-gen target is not invoked as part of a regular build, but
can be run by specifying the target.  If the files are different they
will have to be added and comitted along with the new isa-spec.yaml file.

Still to-do is to have a default target that compare the timestamp of
isa-spec.yaml against that of the pre-generated files.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>